refactor: Migrate `WcPathname` to `WordCount::Pathname`
* 05.wc/lib/wc_methods.rb: Load `05.wc/lib/word_count.rb`.
(#main) <wc_paths>: Use `WordCount::Pathname.new`.

* 05.wc/lib/wc_pathname.rb: Delete file.
<WcPathname>: Migrate to `WordCount::Pathname`.

* 05.wc/lib/word_count.rb: (WordCount::Pathname): New class.
